Fix broken ImageBuf copy constructor from ImageBuf that wraps app buffer (1665)
It was not working, had an assertion.

Remove the assertion, fix the logic to make it work, add a test.

While I was there, I moved a histogram test from imagebuf_test.cpp to
imagebufalgo_test.cpp, it seemed to be more on-topic there.
