+++ src/libOpenImageIO/imagebufalgo_test.cpp
@@ -573,6 +573,54 @@ void test_computePixelStats ()
 
 
 
+// Tests histogram computation.
+void histogram_computation_test ()
+{
+    const int INPUT_WIDTH   = 64;
+    const int INPUT_HEIGHT  = 64;
+    const int INPUT_CHANNEL = 0;
+
+    const int HISTOGRAM_BINS = 256;
+
+    const int SPIKE1 = 51;  // 0.2f in range 0->1 maps to 51 in range 0->255
+    const int SPIKE2 = 128; // 0.5f in range 0->1 maps to 128 in range 0->255
+    const int SPIKE3 = 204; // 0.8f in range 0->1 maps to 204 in range 0->255
+
+    const int SPIKE1_COUNT = INPUT_WIDTH * 8;
+    const int SPIKE2_COUNT = INPUT_WIDTH * 16;
+    const int SPIKE3_COUNT = INPUT_WIDTH * 40;
+
+    // Create input image with three regions with different pixel values.
+    ImageSpec spec (INPUT_WIDTH, INPUT_HEIGHT, 1, TypeDesc::FLOAT);
+    ImageBuf A (spec);
+
+    float value[] = {0.2f};
+    ImageBufAlgo::fill (A, value, ROI(0, INPUT_WIDTH, 0, 8));
+
+    value[0] = 0.5f;
+    ImageBufAlgo::fill (A, value, ROI(0, INPUT_WIDTH, 8, 24));
+
+    value[0] = 0.8f;
+    ImageBufAlgo::fill (A, value, ROI(0, INPUT_WIDTH, 24, 64));
+
+    // Compute A's histogram.
+    std::vector<imagesize_t> hist;
+    ImageBufAlgo::histogram (A, INPUT_CHANNEL, hist, HISTOGRAM_BINS);
+
+    // Does the histogram size equal the number of bins?
+    OIIO_CHECK_EQUAL (hist.size(), (imagesize_t)HISTOGRAM_BINS);
+
+    // Are the histogram values as expected?
+    OIIO_CHECK_EQUAL (hist[SPIKE1], (imagesize_t)SPIKE1_COUNT);
+    OIIO_CHECK_EQUAL (hist[SPIKE2], (imagesize_t)SPIKE2_COUNT);
+    OIIO_CHECK_EQUAL (hist[SPIKE3], (imagesize_t)SPIKE3_COUNT);
+    for (int i = 0; i < HISTOGRAM_BINS; i++)
+        if (i!=SPIKE1 && i!=SPIKE2 && i!=SPIKE3)
+            OIIO_CHECK_EQUAL (hist[i], 0);
+}
+
+
+
 // Test ability to do a maketx directly from an ImageBuf
 void
 test_maketx_from_imagebuf()
@@ -761,6 +809,7 @@ main (int argc, char **argv)
     test_isConstantChannel ();
     test_isMonochrome ();
     test_computePixelStats ();
+    histogram_computation_test ();
     test_maketx_from_imagebuf ();
     test_IBAprep ();
